Protein Dataset & Graph Attention pipeline
Currently the model compiles and iterates just fine, but I haven't
checked it for correctness or accuracy. I should make some changes
later w.r.t evaluation metrics, training flow and so on.

Ideally the protein dataset class can be reused if we're going to
use different models built into Spektral. I *think* it can still
be reused outside of that library because it pretty much returns
a tensorflow dataset.